+++ DockDoor/Views/Intro/Tabs/FirstTimeCongratsTabView.swift
@@ -1,0 +1,12 @@
+import SwiftUI
+
+struct FirstTimeCongratsTabView: View {
+    var nextTab: () -> Void
+    var body: some View {
+        Text("Hello, World!")
+    }
+}
+
+#Preview {
+    FirstTimeCongratsTabView(nextTab: {})
+}
